Added wrapping for x-axis + windSpeed change
I made the snowflakes wrap around the x-axis when they reach one side (sort of like how they wrap around the y axis when they reach the bottom).

I also implemented the motion sensing to augment the windSpeed when detected, which works, but choppily. Will have to find a way to fine-tune it.